refactor(LineService): `getStations()`을 `Sections`로 메소드 이동
