+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -55,4 +55,8 @@ public class Line extends BaseEntity {
     public List<Section> getSections() {
         return sections.get();
     }
+
+    public List<Station> getStations() {
+        return sections.getStations(this);
+    }
 }
